Add files via upload
KMS Superstore Sales Over Time s a corporate style viz, a distinct Superstore information data set plot displaying increases or diminishes in sales, profit, and shipping cost, however, various questions should occur which will require true visuals to exhibit the lengthen to Which Category is Best Selling and Most Profitable? What is the Best Selling and Most Profitable Sub-Category? Which is the Top Selling Sub-Category? Which Customer Segment is Most Profitable? Which is the Preferred Ship Mode? Which Region is the Most Profitable?